test(integration): relocate API-dependent UI tests to tests/integration/
test_get_dataset and test_create_diff_success mock the directly-called view functions
but still reach the running API for secondary calls (/aps, /tasks/active), so they are
integration tests, not unit. Move them to tests/integration/ (marked `integration` by
path) and drop the CI --deselect hack now that the Connexion-3 serving works.

tests/unit is now fully green: 105 passed, coverage 67% > 65% gate.
